add cursor

+++ pager.cpp
@@ -5,6 +5,9 @@
 #include "pager.h"
 
 Pager *pager_open(const char* filename) {
+
+    // Cache for pages
+
     int fd = open(filename, _O_RDWR | _O_CREAT, _S_IWUSR, _S_IRUSR);
 
     if (fd == -1) {
@@ -61,7 +64,7 @@ void pager_flush(Pager *pager, uint32_t page_num, uint32_t size) {
         exit(EXIT_FAILURE);
     }
 
-    ssize_t bytes_written = write(pager->fd, pager->pages[page_num], size);
+    int bytes_written = write(pager->fd, pager->pages[page_num], size);
 
     if (bytes_written == -1) {
         std::cout << "Critical Error: writing to file failed\n";
